moved functions to lib/read-files.js and called from index.js
